Remove pinned projects
